requirements.txt
flask
langchain
chromadb
huggingface-hub
sentence-transformers
transformers
torch
python-dotenv
pytest
requests
gunicorn
